refactor(completionsForms): extract common components
