+++ src/components/forms/TrackerForm/types.ts
@@ -1,4 +1,4 @@
-import Tracker from '../../models/Tracker';
+import Tracker from '../../../models/Tracker';
 
 // Number "duration" and "requiredCompletions.quantity" are used as a string for input control here.
 export type FormValues = Omit<Tracker, 'duration' | 'requiredCompletions'> & {
