+++ src/components/forms/completions/CompletionQuantityTextField.tsx
@@ -1,0 +1,11 @@
+import { TextField } from '@mui/material';
+import styled from '@emotion/styled';
+
+const CompletionQuantityTextField = styled(TextField)`
+  fieldset {
+    border-top-left-radius: 0;
+    border-bottom-left-radius: 0;
+  }
+`;
+
+export default CompletionQuantityTextField;
