+++ src/components/forms/completions/CompletionUnitTextField.tsx
@@ -1,0 +1,11 @@
+import { TextField } from '@mui/material';
+import styled from '@emotion/styled';
+
+const CompletionUnitTextField = styled(TextField)`
+  fieldset {
+    border-top-right-radius: 0;
+    border-bottom-right-radius: 0;
+  }
+`;
+
+export default CompletionUnitTextField;
